+++ docker/entrypoint/node.sh
@@ -13,12 +13,14 @@ if [ "$(opkg list-installed | awk '{ print $1}' | grep -E 'node')" != 'node' ]; 
 	npm install -g http
 
 	## create dir for node apps
-	rmdir /data/node 2>/dev/null
-	mv -f /docker/tmp/node /data
-	find /data/node -print | xargs chown docker:docker
-	find /data/node -type d -print | xargs chmod 755
-	find /data/node -type f -print | xargs chmod 644
-	rmdir /docker/tmp/node 2>/dev/null
+	if [ ! -d /data/node ]; then
+		rm -f /data/node 2>/dev/null
+		mv -f /docker/tmp/node /data
+		find /data/node -print | xargs chown docker:docker
+		find /data/node -type d -print | xargs chmod 755
+		find /data/node -type f -print | xargs chmod 644
+		rmdir /docker/tmp/node 2>/dev/null
+	fi
 fi
 
 ## start node from directory
